Create SmoothCameraMovement
This is a code for a flexible code for a smooth camera movement in unity.
target - The GameObject which the camera needs to follow.
lockTransitionX, lockTransistionY, lockTransitionZ - locks the movement of Camera on the specified Axes (zero to lock the axes, 1 to unlock the axes).
smoothTime - The time the Camera consumes before it reaches the target. A higher value represents a smoother movement than a smaller value.
depth - The z position of the camera.
whenToMove - The distance between the target and the camera, at which the camera starts to move towards the target.